+++ MaxMind.GeoIP2/Exceptions/AddressNotFoundException.cs
@@ -1,6 +1,7 @@
 ﻿#region
 
 using System;
+using System.Runtime.Serialization;
 #endregion
 
 namespace MaxMind.GeoIP2.Exceptions
@@ -9,6 +10,7 @@ namespace MaxMind.GeoIP2.Exceptions
     ///     This exception is thrown when the IP address is not found in the database.
     ///     This generally means that the address was a private or reserved address.
     /// </summary>
+    [Serializable]
     public class AddressNotFoundException : GeoIP2Exception
     {
         /// <summary>
@@ -27,6 +29,15 @@ namespace MaxMind.GeoIP2.Exceptions
         /// <param name="innerException">The inner exception.</param>
         public AddressNotFoundException(string message, Exception innerException)
             : base(message, innerException)
+        {
+        }
+
+        /// <summary>
+        ///     Constructor for deserialization.
+        /// </summary>
+        /// <param name="info"></param>
+        /// <param name="context"></param>
+        protected AddressNotFoundException(SerializationInfo info, StreamingContext context) : base(info, context)
         {
         }
     }