+++ MaxMind.GeoIP2/Exceptions/AuthenticationException.cs
@@ -1,6 +1,7 @@
 ﻿#region
 
 using System;
+using System.Runtime.Serialization;
 #endregion
 
 namespace MaxMind.GeoIP2.Exceptions
@@ -8,6 +9,7 @@ namespace MaxMind.GeoIP2.Exceptions
     /// <summary>
     ///     This exception is thrown when there is an authentication error.
     /// </summary>
+    [Serializable]
     public class AuthenticationException : GeoIP2Exception
     {
         /// <summary>
@@ -26,6 +28,14 @@ namespace MaxMind.GeoIP2.Exceptions
         /// <param name="innerException">The inner exception.</param>
         public AuthenticationException(string message, Exception innerException)
             : base(message, innerException)
+        {
+        }
+        /// <summary>
+        ///     Constructor for deserialization.
+        /// </summary>
+        /// <param name="info"></param>
+        /// <param name="context"></param>
+        protected AuthenticationException(SerializationInfo info, StreamingContext context) : base(info, context)
         {
         }
     }