+++ MaxMind.GeoIP2/Exceptions/GeoIP2Exception.cs
@@ -1,6 +1,7 @@
 ﻿#region
 
 using System;
+using System.Runtime.Serialization;
 #endregion
 
 namespace MaxMind.GeoIP2.Exceptions
@@ -9,6 +10,7 @@ namespace MaxMind.GeoIP2.Exceptions
     ///     This class represents a generic GeoIP2 error. All other exceptions thrown by
     ///     the GeoIP2 API subclass this exception
     /// </summary>
+    [Serializable]
     public class GeoIP2Exception : Exception
     {
         /// <summary>
@@ -26,6 +28,15 @@ namespace MaxMind.GeoIP2.Exceptions
         /// <param name="innerException">The inner exception.</param>
         public GeoIP2Exception(string message, Exception innerException)
             : base(message, innerException)
+        {
+        }
+
+        /// <summary>
+        ///     Constructor for deserialization.
+        /// </summary>
+        /// <param name="info"></param>
+        /// <param name="context"></param>
+        protected GeoIP2Exception(SerializationInfo info, StreamingContext context) : base(info, context)
         {
         }
     }