+++ MaxMind.GeoIP2/Exceptions/HttpException.cs
@@ -3,6 +3,7 @@
 using System;
 using System.IO;
 using System.Net;
+using System.Runtime.Serialization;
 #endregion
 
 namespace MaxMind.GeoIP2.Exceptions
@@ -12,6 +13,7 @@ namespace MaxMind.GeoIP2.Exceptions
     ///     by the web service itself. As such, it is a IOException instead of a
     ///     GeoIP2Exception.
     /// </summary>
+    [Serializable]
     public class HttpException : IOException
     {
         /// <summary>
@@ -36,13 +38,39 @@ namespace MaxMind.GeoIP2.Exceptions
         /// <param name="httpStatus">The HTTP status of the response that caused the exception.</param>
         /// <param name="uri">The URL queried.</param>
         /// <param name="innerException">The underlying exception that caused this one.</param>
-        public HttpException(string message, HttpStatusCode httpStatus, Uri uri, Exception innerException)
+        public HttpException(string message, HttpStatusCode httpStatus, Uri uri, Exception? innerException)
             : base(message, innerException)
         {
             HttpStatus = httpStatus;
 #pragma warning disable IDE0003 // Mono gets confused without 'this'
             this.Uri = uri;
 #pragma warning restore IDE0003
+        }
+
+        /// <summary>
+        ///     Constructor for deserialization.
+        /// </summary>
+        /// <param name="info"></param>
+        /// <param name="context"></param>
+        protected HttpException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+            HttpStatus = (HttpStatusCode) (info.GetValue("MaxMind.GeoIP2.Exceptions.HttpException.HttpStatus", typeof(HttpStatusCode))
+                ?? throw new SerializationException("Unexcepted null HttpStatus value"));
+            Uri = (Uri) (info.GetValue("MaxMind.GeoIP2.Exceptions.HttpException.Uri", typeof(Uri))
+                ?? throw new SerializationException("Unexcepted null Uri value"));
+        }
+
+        /// <summary>
+        ///     Method to serialize data.
+        /// </summary>
+        /// <param name="info"></param>
+        /// <param name="context"></param>
+        public override void GetObjectData(SerializationInfo info, StreamingContext context)
+        {
+            base.GetObjectData(info, context);
+
+            info.AddValue("MaxMind.GeoIP2.Exceptions.HttpException.HttpStatus", HttpStatus, typeof(HttpStatusCode));
+            info.AddValue("MaxMind.GeoIP2.Exceptions.HttpException.Uri", Uri, typeof(Uri));
         }
 
         /// <summary>