+++ MaxMind.GeoIP2/Exceptions/PermissionRequiredException.cs
@@ -1,8 +1,16 @@
-﻿namespace MaxMind.GeoIP2.Exceptions
+﻿#region
+
+using System;
+using System.Runtime.Serialization;
+
+#endregion
+
+namespace MaxMind.GeoIP2.Exceptions
 {
     /// <summary>
     /// This class represents an authentication error.
     /// </summary>
+    [Serializable]
     public class PermissionRequiredException : GeoIP2Exception
     {
         /// <summary>
@@ -20,6 +28,15 @@
         /// <param name="innerException">The underlying exception that caused this one.</param>
         public PermissionRequiredException(string message, System.Exception innerException)
             : base(message, innerException)
+        {
+        }
+
+        /// <summary>
+        ///     Constructor for deserialization.
+        /// </summary>
+        /// <param name="info"></param>
+        /// <param name="context"></param>
+        protected PermissionRequiredException(SerializationInfo info, StreamingContext context) : base(info, context)
         {
         }
     }