+++ MaxMind.GeoIP2/Model/City.cs
@@ -1,8 +1,8 @@
 ﻿#region
 
 using MaxMind.Db;
-using Newtonsoft.Json;
 using System.Collections.Generic;
+using System.Text.Json.Serialization;
 
 #endregion
 
@@ -30,8 +30,8 @@ namespace MaxMind.GeoIP2.Model
         [Constructor]
         public City(int? confidence = null,
             [Parameter("geoname_id")] long? geoNameId = null,
-            IDictionary<string, string>? names = null,
-            IEnumerable<string>? locales = null)
+            IReadOnlyDictionary<string, string>? names = null,
+            IReadOnlyList<string>? locales = null)
             : base(geoNameId, names, locales)
         {
             Confidence = confidence;
@@ -42,7 +42,8 @@ namespace MaxMind.GeoIP2.Model
         ///     is correct. This value is only set when using the Insights
         ///     web service or the Enterprise database.
         /// </summary>
-        [JsonProperty("confidence")]
+        [JsonInclude]
+        [JsonPropertyName("confidence")]
         public int? Confidence { get; internal set; }
     }
 }