+++ MaxMind.GeoIP2/Model/NamedEntity.cs
@@ -1,9 +1,10 @@
 ﻿#region
 
 using MaxMind.Db;
-using Newtonsoft.Json;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
+using System.Text.Json.Serialization;
 
 #endregion
 
@@ -14,22 +15,16 @@ namespace MaxMind.GeoIP2.Model
     /// </summary>
     public abstract class NamedEntity
     {
-        [JsonProperty("names")]
-        private readonly IDictionary<string, string> _names;
-
         /// <summary>
         ///     Constructor
         /// </summary>
         [Constructor]
-        protected NamedEntity(long? geoNameId = null, IDictionary<string, string>? names = null,
-            IEnumerable<string>? locales = null)
+        protected NamedEntity(long? geoNameId = null, IReadOnlyDictionary<string, string>? names = null,
+            IReadOnlyList<string>? locales = null)
         {
-            _names = names != null ? new Dictionary<string, string>(names) : new Dictionary<string, string>();
-            // Unfortunately the existing models incorrectly use an int rather
-            // than a long for the geoname_id. This should be corrected if we
-            // ever do a major version bump.
-            GeoNameId = (int?)geoNameId;
-            Locales = locales != null ? new List<string>(locales) : new List<string> { "en" };
+            Names = names ?? new ReadOnlyDictionary<string, string>(new Dictionary<string, string>());
+            GeoNameId = geoNameId;
+            Locales = locales ?? new List<string> { "en" }.AsReadOnly();
         }
 
         /// <summary>
@@ -40,20 +35,22 @@ namespace MaxMind.GeoIP2.Model
         ///         cred="GeoNameId" />
         ///     or relevant code instead.
         /// </summary>
-        [JsonIgnore]
-        public Dictionary<string, string> Names => new Dictionary<string, string>(_names);
+        [JsonInclude]
+        [JsonPropertyName("names")]
+        public IReadOnlyDictionary<string, string> Names { get; internal set; }
 
         /// <summary>
         ///     The GeoName ID for the city.
         /// </summary>
-        [JsonProperty("geoname_id")]
-        public int? GeoNameId { get; internal set; }
+        [JsonInclude]
+        [JsonPropertyName("geoname_id")]
+        public long? GeoNameId { get; internal set; }
 
         /// <summary>
         ///     Gets or sets the locales specified by the user.
         /// </summary>
         [JsonIgnore]
-        protected internal IEnumerable<string> Locales { get; set; }
+        protected internal IReadOnlyList<string> Locales { get; set; }
 
         /// <summary>
         ///     The name of the city based on the locales list passed to the
@@ -68,9 +65,8 @@ namespace MaxMind.GeoIP2.Model
         {
             get
             {
-                var names = _names;
-                var locale = Locales.FirstOrDefault(l => names.ContainsKey(l));
-                return locale == null ? null : names[locale];
+                var locale = Locales.FirstOrDefault(l => Names.ContainsKey(l));
+                return locale == null ? null : Names[locale];
             }
         }