+++ MaxMind.GeoIP2/Responses/ConnectionTypeResponse.cs
@@ -1,7 +1,7 @@
 ﻿#region
 
 using MaxMind.Db;
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 
 #endregion
 
@@ -37,7 +37,8 @@ namespace MaxMind.GeoIP2.Responses
         /// <summary>
         ///     The connection type of the IP address.
         /// </summary>
-        [JsonProperty("connection_type")]
+        [JsonInclude]
+        [JsonPropertyName("connection_type")]
         public string? ConnectionType { get; internal set; }
 
         /// <summary>
@@ -47,7 +48,8 @@ namespace MaxMind.GeoIP2.Responses
         ///     on. If the system is behind a NAT, this may differ from the IP
         ///     address locally assigned to it.
         /// </summary>
-        [JsonProperty("ip_address")]
+        [JsonInclude]
+        [JsonPropertyName("ip_address")]
         public string? IPAddress { get; internal set; }
 
         /// <summary>
@@ -55,7 +57,8 @@ namespace MaxMind.GeoIP2.Responses
         ///     the largest network where all of the fields besides
         ///     <c>IPAddress</c> have the same value.
         /// </summary>
-        [JsonProperty("network")]
+        [JsonInclude]
+        [JsonPropertyName("network")]
         public Network? Network { get; internal set; }
     }
 }