+++ MaxMind.GeoIP2/Responses/IspResponse.cs
@@ -1,7 +1,7 @@
 ﻿#region
 
 using MaxMind.Db;
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 
 #endregion
 
@@ -48,7 +48,8 @@ namespace MaxMind.GeoIP2.Responses
         ///     </a>
         ///     associated with the IP address.
         /// </summary>
-        [JsonProperty("autonomous_system_number")]
+        [JsonInclude]
+        [JsonPropertyName("autonomous_system_number")]
         public long? AutonomousSystemNumber { get; internal set; }
 
         /// <summary>
@@ -59,19 +60,22 @@ namespace MaxMind.GeoIP2.Responses
         ///     </a>
         ///     for the IP address.
         /// </summary>
-        [JsonProperty("autonomous_system_organization")]
+        [JsonInclude]
+        [JsonPropertyName("autonomous_system_organization")]
         public string? AutonomousSystemOrganization { get; internal set; }
 
         /// <summary>
         ///     The name of the ISP associated with the IP address.
         /// </summary>
-        [JsonProperty("isp")]
+        [JsonInclude]
+        [JsonPropertyName("isp")]
         public string? Isp { get; internal set; }
 
         /// <summary>
         ///     The name of the organization associated with the IP address.
         /// </summary>
-        [JsonProperty("organization")]
+        [JsonInclude]
+        [JsonPropertyName("organization")]
         public string? Organization { get; internal set; }
 
         /// <summary>
@@ -81,7 +85,8 @@ namespace MaxMind.GeoIP2.Responses
         ///     on. If the system is behind a NAT, this may differ from the IP
         ///     address locally assigned to it.
         /// </summary>
-        [JsonProperty("ip_address")]
+        [JsonInclude]
+        [JsonPropertyName("ip_address")]
         public string? IPAddress { get; internal set; }
 
         /// <summary>
@@ -89,7 +94,8 @@ namespace MaxMind.GeoIP2.Responses
         ///     the largest network where all of the fields besides
         ///     <c>IPAddress</c> have the same value.
         /// </summary>
-        [JsonProperty("network")]
+        [JsonInclude]
+        [JsonPropertyName("network")]
         public Network? Network { get; internal set; }
     }
 }